+++ sites/partners/lib/formatApplicationData.ts
@@ -7,6 +7,7 @@ import {
   ApplicationStatus,
   AddressUpdate,
   HouseholdMember,
+  Program,
 } from "@bloom-housing/backend-core/types"
 
 import {
@@ -14,6 +15,11 @@ import {
   mapPreferencesToApi,
   mapApiToPreferencesForm,
 } from "@bloom-housing/ui-components"
+import {
+  fieldGroupObjectToArray,
+  mapProgramsToApi,
+  mapApiToProgramsPaperForm,
+} from "@bloom-housing/shared-helpers"
 import {
   FormTypes,
   YesNoAnswer,
@@ -53,11 +59,18 @@ interface FormData extends FormTypes {
   submissionType: ApplicationSubmissionType
 }
 
+type mapFormToApiProps = {
+  data: FormData
+  listingId: string
+  editMode: boolean
+  programs: Program[]
+}
+
 /*
   Format data which comes from react-hook-form into correct API format.
 */
 
-export const mapFormToApi = (data: FormData, listingId: string, editMode: boolean) => {
+export const mapFormToApi = ({ data, listingId, editMode, programs }: mapFormToApiProps) => {
   const language: Language | null = data.application?.language ? data.application?.language : null
 
   const submissionDate: Date | null = (() => {
@@ -112,6 +125,14 @@ export const mapFormToApi = (data: FormData, listingId: string, editMode: boolea
   })()
 
   const preferences = mapPreferencesToApi(data)
+  const programsForm = Object.entries(data.application.programs).reduce((acc, curr) => {
+    if (curr[1]) {
+      Object.assign(acc, { [curr[0]]: curr[1] })
+    }
+    return acc
+  }, {})
+
+  const programsData = mapProgramsToApi(programs, programsForm)
 
   // additional phone
   const {
@@ -122,13 +143,17 @@ export const mapFormToApi = (data: FormData, listingId: string, editMode: boolea
     contactPreferences,
     sendMailToMailingAddress,
     accessibility,
-    demographics,
   } = data.application
 
   const additionalPhone = !additionalPhoneNumberData
   const additionalPhoneNumberType = additionalPhoneNumberTypeData
     ? additionalPhoneNumberTypeData
     : null
+
+  const demographics = {
+    ...data.application.demographics,
+    race: fieldGroupObjectToArray(data, "race"),
+  }
 
   const mailingAddress = getAddress(sendMailToMailingAddress, mailingAddressData)
 
@@ -179,6 +204,7 @@ export const mapFormToApi = (data: FormData, listingId: string, editMode: boolea
     householdExpectingChanges,
     householdStudent,
     preferences,
+    programs: programsData,
     income,
     incomePeriod,
     incomeVouchers,
@@ -252,6 +278,7 @@ export const mapApiToForm = (applicationData: ApplicationUpdate) => {
   const phoneNumber = applicationData.applicant.phoneNumber
 
   const preferences = mapApiToPreferencesForm(applicationData.preferences)
+  const programs = mapApiToProgramsPaperForm(applicationData.programs)
 
   const application: ApplicationTypes = (() => {
     const {
@@ -300,6 +327,7 @@ export const mapApiToForm = (applicationData: ApplicationUpdate) => {
       demographics,
       acceptedTerms,
       alternateContact,
+      programs,
     }
 
     return result