+++ sites/public/cypress/mockData/applicationData.ts
@@ -30,7 +30,7 @@ export const preferredUnitCheckboxesOrder = [
 ]
 
 export const howDidYouHearCheckboxesOrder = [
-  "alamedaCountyHCDWebsite",
+  "jurisdictionWebsite",
   "developerWebsite",
   "flyer",
   "emailAlert",
@@ -39,6 +39,25 @@ export const howDidYouHearCheckboxesOrder = [
   "radioAd",
   "busAd",
   "other",
+]
+
+export const raceCheckboxesOrder = [
+  "americanIndianAlaskanNative",
+  "asian",
+  "asian-asianIndian",
+  "asian-chinese",
+  "asian-filipino",
+  "asian-korean",
+  "asian-vietnamese",
+  "asian-otherAsian",
+  "blackAfricanAmerican",
+  "nativeHawaiianOtherPacificIslander",
+  "nativeHawaiianOtherPacificIslander-nativeHawaiian",
+  "nativeHawaiianOtherPacificIslander-guamanianOrChamorro",
+  "nativeHawaiianOtherPacificIslander-samoan",
+  "white",
+  "otherMultiracial",
+  "declineToRespond",
 ]
 
 export const coliseumApplication: Application = {
@@ -180,11 +199,11 @@ export const coliseumApplication: Application = {
   ],
   demographics: {
     ...idDefaults,
-    race: "white",
+    race: ["race-white"],
     ethnicity: "hispanicLatino",
-    gender: "female",
-    sexualOrientation: "bisexual",
-    howDidYouHear: ["alamedaCountyHCDWebsite", "developerWebsite"],
+    gender: "",
+    sexualOrientation: "",
+    howDidYouHear: ["jurisdictionWebsite", "developerWebsite"],
   },
   preferences: [
     {
@@ -243,6 +262,44 @@ export const coliseumApplication: Application = {
         },
         {
           key: "doNotConsider",
+          checked: false,
+        },
+      ],
+    },
+  ],
+  programs: [
+    {
+      claimed: true,
+      key: "servedInMilitary",
+      options: [
+        {
+          key: "servedInMilitary",
+          checked: true,
+        },
+        {
+          key: "doNotConsider",
+          checked: false,
+        },
+        {
+          key: "preferNotToSay",
+          checked: false,
+        },
+      ],
+    },
+    {
+      key: "tay",
+      claimed: true,
+      options: [
+        {
+          key: "tay",
+          checked: true,
+        },
+        {
+          key: "doNotConsider",
+          checked: false,
+        },
+        {
+          key: "preferNotToSay",
           checked: false,
         },
       ],
@@ -355,8 +412,87 @@ export const minimalDataApplication: Application = {
   demographics: {
     ...idDefaults,
     howDidYouHear: [],
+    race: [],
   },
   preferences: [],
+  programs: [
+    {
+      claimed: false,
+      key: "servedInMilitary",
+      options: [
+        {
+          key: "servedInMilitary",
+          checked: false,
+        },
+        {
+          key: "doNotConsider",
+          checked: false,
+        },
+        {
+          key: "preferNotToSay",
+          checked: false,
+        },
+      ],
+    },
+    {
+      key: "tay",
+      claimed: false,
+      options: [
+        {
+          key: "tay",
+          checked: false,
+        },
+        {
+          key: "doNotConsider",
+          checked: false,
+        },
+        {
+          key: "preferNotToSay",
+          checked: false,
+        },
+      ],
+    },
+    {
+      key: "disabilityOrMentalIllness",
+      claimed: false,
+      options: [
+        {
+          key: "disabilityOrMentalIllness",
+          checked: false,
+        },
+        {
+          key: "doNotConsider",
+          checked: false,
+        },
+        {
+          key: "preferNotToSay",
+          checked: false,
+        },
+      ],
+    },
+    {
+      key: "housingSituation",
+      claimed: false,
+      options: [
+        {
+          key: "notPermanent",
+          checked: false,
+        },
+        {
+          key: "homeless",
+          checked: false,
+        },
+        {
+          key: "doNotConsider",
+          checked: false,
+        },
+        {
+          key: "preferNotToSay",
+          checked: false,
+        },
+      ],
+    },
+  ],
   confirmationCode: "",
   status: ApplicationStatus.draft,
   submissionType: ApplicationSubmissionType.electronical,
@@ -376,6 +512,7 @@ export const applicationStepOrder = [
   { name: "addMembers", route: "/applications/household/add-members" },
   { name: "preferredUnitSize", route: "/applications/household/preferred-units" },
   { name: "adaHouseholdMembers", route: "/applications/household/ada" },
+  { name: "programs", route: "/applications/household/programs" },
   { name: "householdExpectingChanges", route: "/applications/household/changes" },
   { name: "householdStudent", route: "/applications/household/student" },
   { name: "vouchersSubsidies", route: "/applications/financial/vouchers" },